refactor : ChamponshipForm svg 분리했습니다 커스텀훅 분리했습니다 뷰로직 분리했습니다
