+++ src/1_Page/ChampionshipEdit/lib/useManageSearchParam.ts
@@ -1,4 +1,5 @@
 import { useNavigate, useParams } from "react-router-dom";
+import { useAuthStore } from "../../../4_Shared/lib/useMyInfo";
 
 const useManageSearchParam = (): {
   isEditMode: boolean;
@@ -14,6 +15,12 @@ const useManageSearchParam = (): {
 
   const isEditMode = mode === "edit";
   const parsedCommunityIdx = parseInt(communityIdx || "0", 10);
+
+  const roleIdx = useAuthStore((state) => state.communityRoleIdx);
+  if (roleIdx !== 0) {
+    alert("권한이 없습니다.");
+    navigate("/");
+  }
   return { isEditMode, communityIdx: parsedCommunityIdx };
 };
