+++ src/1_Page/ChampionshipEdit/model/type.d.ts
@@ -1,0 +1,5 @@
+type UseChanpionshipFormProps = {
+  isEditMode: boolean;
+  communityIdx: number;
+  setActiveTab: (tab: ChampionshipEditTab) => void;
+};
